# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminSessionInput {
  /** @minLength 1 */
  idToken: string;
}

export interface AdminIdentity {
  email: string;
}

export interface AdminSession {
  token: string;
  admin: AdminIdentity;
}

export interface ModerationSummary {
  pendingReports: number;
  reportsToday: number;
  actionedReports: number;
  removedUsers: number;
}

export type ModerationUserModerationStatus = typeof ModerationUserModerationStatus[keyof typeof ModerationUserModerationStatus];


export const ModerationUserModerationStatus = {
  active: 'active',
  removed: 'removed',
} as const;

export interface ModerationUser {
  id: string;
  firstName: string;
  email: string;
  /** @nullable */
  university?: string | null;
  /** @nullable */
  profilePicture?: string | null;
  moderationStatus: ModerationUserModerationStatus;
}

export type ModerationReportStatus = typeof ModerationReportStatus[keyof typeof ModerationReportStatus];


export const ModerationReportStatus = {
  pending: 'pending',
  reviewed: 'reviewed',
  dismissed: 'dismissed',
  actioned: 'actioned',
} as const;

export interface ModerationReport {
  id: string;
  category: string;
  details: string;
  status: ModerationReportStatus;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  /** @nullable */
  reviewedAt?: string | null;
  /** @nullable */
  reviewedBy?: string | null;
  reporter: ModerationUser;
  reported: ModerationUser;
}

export type ReportResolutionInputStatus = typeof ReportResolutionInputStatus[keyof typeof ReportResolutionInputStatus];


export const ReportResolutionInputStatus = {
  reviewed: 'reviewed',
  dismissed: 'dismissed',
  actioned: 'actioned',
} as const;

export interface ReportResolutionInput {
  status: ReportResolutionInputStatus;
  /** @maxLength 1000 */
  adminNote?: string;
}

export interface UserRemovalInput {
  /**
     * @minLength 3
     * @maxLength 1000
     */
  reason: string;
  reportId?: string;
}

export interface ModerationActionResult {
  success: boolean;
  userId: string;
}

export type ListModerationReportsParams = {
status?: ListModerationReportsStatus;
};

export type ListModerationReportsStatus = typeof ListModerationReportsStatus[keyof typeof ListModerationReportsStatus];


export const ListModerationReportsStatus = {
  pending: 'pending',
  reviewed: 'reviewed',
  dismissed: 'dismissed',
  actioned: 'actioned',
} as const;