# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: admin
    description: UniLink moderation operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /admin/session:
    post:
      operationId: createAdminSession
      tags: [admin]
      summary: Authenticate a UniLink administrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSessionInput"
      responses:
        "200":
          description: Authenticated administrator session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "403":
          description: Not an authorized administrator
  /admin/summary:
    get:
      operationId: getModerationSummary
      tags: [admin]
      summary: Get moderation queue summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Moderation summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationSummary"
  /admin/reports:
    get:
      operationId: listModerationReports
      tags: [admin]
      summary: List user reports
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, reviewed, dismissed, actioned]
      responses:
        "200":
          description: Report queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModerationReport"
  /admin/reports/{reportId}:
    patch:
      operationId: resolveModerationReport
      tags: [admin]
      summary: Resolve a report
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reportId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportResolutionInput"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationReport"
  /admin/users/{userId}/remove:
    post:
      operationId: removeAbusiveUser
      tags: [admin]
      summary: Remove an abusive user from UniLink
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRemovalInput"
      responses:
        "200":
          description: User removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationActionResult"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AdminSessionInput:
      type: object
      required: [idToken]
      properties:
        idToken:
          type: string
          minLength: 1
    AdminIdentity:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    AdminSession:
      type: object
      required: [token, admin]
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/AdminIdentity"
    ModerationSummary:
      type: object
      required: [pendingReports, reportsToday, actionedReports, removedUsers]
      properties:
        pendingReports:
          type: integer
        reportsToday:
          type: integer
        actionedReports:
          type: integer
        removedUsers:
          type: integer
    ModerationUser:
      type: object
      required: [id, firstName, email, moderationStatus]
      properties:
        id:
          type: string
        firstName:
          type: string
        email:
          type: string
          format: email
        university:
          type: [string, "null"]
        profilePicture:
          type: [string, "null"]
        moderationStatus:
          type: string
          enum: [active, removed]
    ModerationReport:
      type: object
      required:
        [id, category, details, status, createdAt, reporter, reported]
      properties:
        id:
          type: string
        category:
          type: string
        details:
          type: string
        status:
          type: string
          enum: [pending, reviewed, dismissed, actioned]
        adminNote:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        reviewedAt:
          type: [string, "null"]
          format: date-time
        reviewedBy:
          type: [string, "null"]
        reporter:
          $ref: "#/components/schemas/ModerationUser"
        reported:
          $ref: "#/components/schemas/ModerationUser"
    ReportResolutionInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [reviewed, dismissed, actioned]
        adminNote:
          type: string
          maxLength: 1000
    UserRemovalInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 3
          maxLength: 1000
        reportId:
          type: string
    ModerationActionResult:
      type: object
      required: [success, userId]
      properties:
        success:
          type: boolean
        userId:
          type: string