# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminIdentity } from './adminIdentity';

export interface AdminSession {
  token: string;
  admin: AdminIdentity;
}