# lib/api-zod/src/generated/types/moderationReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ModerationReportStatus } from './moderationReportStatus';
import type { ModerationUser } from './moderationUser';

export interface ModerationReport {
  id: string;
  category: string;
  details: string;
  status: ModerationReportStatus;
  /** @nullable */
  adminNote?: string | null;
  createdAt: Date;
  /** @nullable */
  reviewedAt?: Date | null;
  /** @nullable */
  reviewedBy?: string | null;
  reporter: ModerationUser;
  reported: ModerationUser;
}