# lib/api-zod/src/generated/types/moderationUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ModerationUserModerationStatus } from './moderationUserModerationStatus';

export interface ModerationUser {
  id: string;
  firstName: string;
  email: string;
  /** @nullable */
  university?: string | null;
  /** @nullable */
  profilePicture?: string | null;
  moderationStatus: ModerationUserModerationStatus;
}