# lib/api-zod/src/generated/types/reportResolutionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportResolutionInputStatus } from './reportResolutionInputStatus';

export interface ReportResolutionInput {
  status: ReportResolutionInputStatus;
  /** @maxLength 1000 */
  adminNote?: string;
}