# lib/api-zod/src/generated/types/userRemovalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserRemovalInput {
  /**
     * @minLength 3
     * @maxLength 1000
     */
  reason: string;
  reportId?: string;
}